use long instead of ssize_t which does not exist with MSVC

+++ example/api2-samples/api2-decode-audio.cpp
@@ -31,7 +31,7 @@ int main(int argc, char **argv)
 
     string uri {argv[1]};
 
-    ssize_t      audioStream = -1;
+    long      audioStream = -1;
     AudioDecoderContext adec;
     Stream      ast;
     error_code   ec;
